test/inspector/debugger/wasm-jspi-async-stack-expected.txt

Test async stack traces with wasm jspi

Running test: testCatchPrediction
Testing catch prediction through JSPI throwing from testSyncThrow to testCatch
Debugger paused on uncaught exception
testSyncThrow (test.js:56:2)
js_func (test.js:24:17)
$wrappedWasm (wasm://wasm/bbce4a62:0:76)
$threeTimes (wasm://wasm/bbce4a62:0:85)

console: caught: fail
Returned result {"result":{"type":"undefined"}}
PREDICTION MISMATCH: predicted uncaught=true actual uncaught=false

Testing catch prediction through JSPI throwing from testAsyncThrow to testCatch
Debugger paused on caught exception
testAsyncThrow (test.js:61:2)
-- await --
js_func (test.js:24:17)
$wrappedWasm (wasm://wasm/bbce4a62:0:76)
$threeTimes (wasm://wasm/bbce4a62:0:85)

console: caught: fail
Returned result {"result":{"type":"undefined"}}

Testing catch prediction through JSPI throwing from testSyncThrow to testDontCatch
Debugger paused on uncaught exception
testSyncThrow (test.js:56:2)
js_func (test.js:24:17)
$wrappedWasm (wasm://wasm/bbce4a62:0:76)
$threeTimes (wasm://wasm/bbce4a62:0:85)

Returned result {"result":{"type":"undefined"}}
Exception was not caught

Testing catch prediction through JSPI throwing from testAsyncThrow to testDontCatch
Debugger paused on uncaught exception
testAsyncThrow (test.js:61:2)
-- await --
js_func (test.js:24:17)
$wrappedWasm (wasm://wasm/bbce4a62:0:76)
$threeTimes (wasm://wasm/bbce4a62:0:85)

Returned result {"result":{"type":"undefined"}}
Exception was not caught
